refactor(react): Move helpers out of components directory
