+++ react-onsenui/src/baseDialog.jsx
