+++ react-onsenui/src/components/Dialog.jsx
@@ -1,8 +1,8 @@
 import PropTypes from 'prop-types';
 import 'onsenui/esm/elements/ons-dialog';
 
-import onsCustomElement from './onsCustomElement';
-import baseDialog from './baseDialog';
+import onsCustomElement from '../onsCustomElement';
+import baseDialog from '../baseDialog';
 
 const deprecated = {
   onCancel: 'onDialogCancel',
