+++ react-onsenui/src/inputProps.jsx
